+++ src/javascript/binary/websocket_pages/mb_trade/mb_process.js
@@ -8,6 +8,7 @@ var TradingAnalysis    = require('../trade/analysis').TradingAnalysis;
 var japanese_client    = require('../../common_functions/country_base').japanese_client;
 var displayUnderlyings = require('../trade/common').displayUnderlyings;
 var showFormOverlay    = require('../trade/common').showFormOverlay;
+var processForgetTicks = require('../trade/process').processForgetTicks;
 
 var MBProcess = (function() {
     var market_status = '',
