+++ src/javascript/binary/websocket_pages/mb_trade/mb_tradepage.js
@@ -3,14 +3,14 @@ var MBDisplayCurrencies = require('./mb_currency').MBDisplayCurrencies;
 var MBTradingEvents     = require('./mb_event').MBTradingEvents;
 var MBMessage           = require('./mb_message').MBMessage;
 var MBSymbols           = require('./mb_symbols').MBSymbols;
+var TradingAnalysis     = require('../trade/analysis').TradingAnalysis;
+var JapanPortfolio      = require('../../../binary_japan/trade_japan/portfolio').JapanPortfolio;
+var State               = require('../../base/storage').State;
+var Content             = require('../../common_functions/content').Content;
 var MBProcess           = require('./mb_process').MBProcess;
 var MBNotifications     = require('./mb_notifications').MBNotifications;
 var MBPrice             = require('./mb_price').MBPrice;
-var TradingAnalysis     = require('../trade/analysis').TradingAnalysis;
 var chartFrameCleanup   = require('../trade/common').chartFrameCleanup;
-var State               = require('../../base/storage').State;
-var Content             = require('../../common_functions/content').Content;
-var JapanPortfolio      = require('../../../binary_japan/trade_japan/portfolio').JapanPortfolio;
 
 var MBTradePage = (function(){
