+++ src/javascript/binary/websocket_pages/trade/event.js
@@ -4,8 +4,13 @@ var Contract                   = require('./contract').Contract;
 var Defaults                   = require('./defaults').Defaults;
 var Durations                  = require('./duration').Durations;
 var Price                      = require('./price').Price;
-var StartDates                 = require('./starttime').StartDates;
+var getStartDatenode           = require('./common_independent').getStartDatenode;
 var Tick                       = require('./tick').Tick;
+var processMarket              = require('./process').processMarket;
+var processContractForm        = require('./process').processContractForm;
+var processForgetTicks         = require('./process').processForgetTicks;
+var onExpiryTypeChange         = require('./process').onExpiryTypeChange;
+var onDurationUnitChange       = require('./process').onDurationUnitChange;
 var onlyNumericOnKeypress      = require('../../common_functions/event_handler').onlyNumericOnKeypress;
 var onlyNumericColonOnKeypress = require('../../common_functions/event_handler').onlyNumericColonOnKeypress;
 var moment                     = require('moment');
@@ -34,52 +39,6 @@ var dateValueChanged           = require('../../common_functions/common_function
  */
 var TradingEvents = (function () {
     'use strict';
-
-    var onExpiryTypeChange = function(value){
-        if(!value || !$('#expiry_type').find('option[value='+value+']').length){
-            value = 'duration';
-        }
-        $('#expiry_type').val(value);
-
-        var make_price_request = 0;
-        if(value === 'endtime'){
-            Durations.displayEndTime();
-            if(Defaults.get('expiry_date')){
-                Durations.selectEndDate(moment(Defaults.get('expiry_date')));
-                make_price_request = -1;
-            }
-            Defaults.remove('duration_units', 'duration_amount');
-        } else {
-            StartDates.enable();
-            Durations.display();
-            if(Defaults.get('duration_units')){
-                TradingEvents.onDurationUnitChange(Defaults.get('duration_units'));
-            }
-            var duration_amount = Defaults.get('duration_amount');
-            if(duration_amount && duration_amount > $('#duration_minimum').text()){
-                $('#duration_amount').val(duration_amount);
-            }
-            make_price_request = 1;
-            Defaults.remove('expiry_date', 'expiry_time', 'end_date');
-            Durations.validateMinDurationAmount();
-        }
-
-        return make_price_request;
-    };
-
-    var onDurationUnitChange = function(value){
-        if(!value || !$('#duration_units').find('option[value='+value+']').length){
-            return 0;
-        }
-
-        $('#duration_units').val(value);
-        Defaults.set('duration_units', value);
-
-        Durations.select_unit(value);
-        Durations.populate();
-
-        return 1;
-    };
 
     var initiate = function () {
         /*
@@ -180,7 +139,7 @@ var TradingEvents = (function () {
             }
             Defaults.set('duration_amount', e.target.value);
             Durations.select_amount(e.target.value);
-            processPriceRequest();
+            Price.processPriceRequest();
             submitForm(document.getElementById('websocket_form'));
         }
         var durationAmountElement = document.getElementById('duration_amount'),
@@ -212,7 +171,7 @@ var TradingEvents = (function () {
             expiryTypeElement.addEventListener('change', function(e) {
                 Defaults.set('expiry_type', e.target.value);
                 onExpiryTypeChange(e.target.value);
-                if (expiryTypeElement.value !== 'endtime') processPriceRequest();
+                if (expiryTypeElement.value !== 'endtime') Price.processPriceRequest();
             });
         }
 
@@ -224,7 +183,7 @@ var TradingEvents = (function () {
             durationUnitElement.addEventListener('change', function (e) {
                 Defaults.remove('barrier', 'barrier_high', 'barrier_low');
                 onDurationUnitChange(e.target.value);
-                processPriceRequest();
+                Price.processPriceRequest();
             });
         }
 
@@ -255,7 +214,7 @@ var TradingEvents = (function () {
                     }
                     if (timeIsValid($('#expiry_time'))) {
                         Durations.setTime(endTimeElement.value);
-                        processPriceRequest();
+                        Price.processPriceRequest();
                     }
                 });
         }
@@ -273,7 +232,7 @@ var TradingEvents = (function () {
                     e.target.value = parseFloat(e.target.value).toFixed(2);
                 }
                 Defaults.set('amount', e.target.value);
-                processPriceRequest();
+                Price.processPriceRequest();
                 submitForm(document.getElementById('websocket_form'));
             }));
         }
@@ -283,13 +242,13 @@ var TradingEvents = (function () {
          * whether start time is forward starting or not and request
          * new price
          */
-        var dateStartElement = StartDates.node();
+        var dateStartElement = getStartDatenode();
         if (dateStartElement) {
             dateStartElement.addEventListener('change', function (e) {
                 Defaults.set('date_start', e.target.value);
                 var r = Durations.onStartDateChange(e.target.value);
                 if(r>=0){
-                    processPriceRequest();
+                    Price.processPriceRequest();
                 }
             });
         }
@@ -302,7 +261,7 @@ var TradingEvents = (function () {
         if (amountTypeElement) {
             amountTypeElement.addEventListener('change', function (e) {
                 Defaults.set('amount_type', e.target.value);
-                processPriceRequest();
+                Price.processPriceRequest();
             });
         }
 
@@ -345,7 +304,7 @@ var TradingEvents = (function () {
                 if (stopTypeDollarLabel && isVisible(stopTypeDollarLabel)) {
                     stopTypeDollarLabel.textContent = e.target.value;
                 }
-                processPriceRequest();
+                Price.processPriceRequest();
             });
         }
 
@@ -371,7 +330,7 @@ var TradingEvents = (function () {
                 if (id && askPrice) {
                     BinarySocket.send(params);
                     Price.incrFormId();
-                    processForgetProposals();
+                    Price.processForgetProposals();
                 }
             }
         });
@@ -384,7 +343,7 @@ var TradingEvents = (function () {
                 e.preventDefault();
                 document.getElementById('contract_confirmation_container').style.display = 'none';
                 document.getElementById('contracts_list').style.display = 'flex';
-                processPriceRequest();
+                Price.processPriceRequest();
             }
         });
 
@@ -396,7 +355,7 @@ var TradingEvents = (function () {
             barrierElement.addEventListener('input', debounce( function (e) {
                 Barriers.validateBarrier();
                 Defaults.set('barrier', e.target.value);
-                processPriceRequest();
+                Price.processPriceRequest();
                 submitForm(document.getElementById('websocket_form'));
             }, 1000));
         }
@@ -408,7 +367,7 @@ var TradingEvents = (function () {
         if (lowBarrierElement) {
             lowBarrierElement.addEventListener('input', debounce( function (e) {
                 Defaults.set('barrier_low', e.target.value);
-                processPriceRequest();
+                Price.processPriceRequest();
                 submitForm(document.getElementById('websocket_form'));
             }));
         }
@@ -420,7 +379,7 @@ var TradingEvents = (function () {
         if (highBarrierElement) {
             highBarrierElement.addEventListener('input', debounce( function (e) {
                 Defaults.set('barrier_high', e.target.value);
-                processPriceRequest();
+                Price.processPriceRequest();
                 submitForm(document.getElementById('websocket_form'));
             }));
         }
@@ -433,7 +392,7 @@ var TradingEvents = (function () {
 
             predictionElement.addEventListener('change', debounce( function (e) {
                 Defaults.set('prediction', e.target.value);
-                processPriceRequest();
+                Price.processPriceRequest();
                 submitForm(document.getElementById('websocket_form'));
             }));
         }
@@ -448,7 +407,7 @@ var TradingEvents = (function () {
                     e.target.value = parseFloat(e.target.value).toFixed(2);
                 }
                 Defaults.set('amount_per_point', e.target.value);
-                processPriceRequest();
+                Price.processPriceRequest();
                 submitForm(document.getElementById('websocket_form'));
             }));
         }
@@ -458,7 +417,7 @@ var TradingEvents = (function () {
          */
         var stopTypeEvent = function (e) {
             Defaults.set('stop_type', e.target.value);
-            processPriceRequest();
+            Price.processPriceRequest();
         };
 
         var stopTypeElement = document.querySelectorAll('input[name="stop_type"]');
@@ -478,7 +437,7 @@ var TradingEvents = (function () {
                     e.target.value = parseFloat(e.target.value).toFixed(2);
                 }
                 Defaults.set('stop_loss', e.target.value);
-                processPriceRequest();
+                Price.processPriceRequest();
                 submitForm(document.getElementById('websocket_form'));
             }));
         }
@@ -493,7 +452,7 @@ var TradingEvents = (function () {
                     e.target.value = parseFloat(e.target.value).toFixed(2);
                 }
                 Defaults.set('stop_profit', e.target.value);
-                processPriceRequest();
+                Price.processPriceRequest();
                 submitForm(document.getElementById('websocket_form'));
             }));
         }
@@ -537,8 +496,6 @@ var TradingEvents = (function () {
 
     return {
         init: initiate,
-        onExpiryTypeChange  : onExpiryTypeChange,
-        onDurationUnitChange: onDurationUnitChange
     };
 })();
