+++ src/javascript/binary/websocket_pages/trade/price.js
@@ -1,5 +1,4 @@
-var Durations                  = require('./duration').Durations;
-var StartDates                 = require('./starttime').StartDates;
+var getStartDatenode           = require('./common_independent').getStartDatenode;
 var Content                    = require('../../common_functions/content').Content;
 var format_money               = require('../../common_functions/currency_to_symbol').format_money;
 var moment                     = require('moment');
@@ -7,7 +6,11 @@ var contractTypeDisplayMapping = require('./common').contractTypeDisplayMapping;
 var resetPriceMovement         = require('./common').resetPriceMovement;
 var displayCommentPrice        = require('./common').displayCommentPrice;
 var displayCommentSpreads      = require('./common').displayCommentSpreads;
+var showPriceOverlay           = require('./common').showPriceOverlay;
 var displayPriceMovement       = require('./common_independent').displayPriceMovement;
+var getTradingTimes            = require('./common_independent').getTradingTimes;
+var Contract                   = require('./contract').Contract;
+var Defaults                   = require('./defaults').Defaults;
 var isVisible                  = require('../../common_functions/common_functions').isVisible;
 
 /*
@@ -24,8 +27,6 @@ var isVisible                  = require('../../common_functions/common_function
  * `Price.display()` to display the price details returned from server
  */
 var Price = (function() {
-    'use strict';
-
     var typeDisplayIdMapping = {},
         form_id = 0;
 
@@ -39,7 +40,7 @@ var Price = (function() {
             amountType = document.getElementById('amount_type'),
             currency = document.getElementById('currency'),
             payout = document.getElementById('amount'),
-            startTime = StartDates.node(),
+            startTime = getStartDatenode(),
             expiryType = document.getElementById('expiry_type'),
             duration = document.getElementById('duration_amount'),
             durationUnit = document.getElementById('duration_units'),
@@ -82,9 +83,9 @@ var Price = (function() {
             proposal['duration_unit'] = durationUnit.value;
         } else if (expiryType && isVisible(expiryType) && expiryType.value === 'endtime') {
             var endDate2 = endDate.getAttribute('data-value');
-            var endTime2 = Durations.getTime();
+            var endTime2 = Defaults.get('expiry_time');
             if (!endTime2) {
-                var trading_times = Durations.trading_times();
+                var trading_times = getTradingTimes();
                 if (trading_times.hasOwnProperty(endDate2) && typeof trading_times[endDate2][underlying.value] === 'object' && trading_times[endDate2][underlying.value].length && trading_times[endDate2][underlying.value][0] !== '--') {
                     if( trading_times[endDate2][underlying.value].length>1)
                         endTime2 = trading_times[endDate2][underlying.value][1];
@@ -275,6 +276,57 @@ var Price = (function() {
         form_id = 0;
     };
 
+    /*
+     * Function to request for cancelling the current price proposal
+     */
+    function processForgetProposals() {
+        'use strict';
+        showPriceOverlay();
+        BinarySocket.send({
+            forget_all: "proposal"
+        });
+        Price.clearMapping();
+    }
+
+    /*
+     * Function to process and calculate price based on current form
+     * parameters or change in form parameters
+     */
+    function processPriceRequest() {
+        'use strict';
+
+        Price.incrFormId();
+        processForgetProposals();
+        showPriceOverlay();
+        var types = Contract.contractType()[Contract.form()];
+        if (Contract.form() === 'digits') {
+            switch (sessionStorage.getItem('formname')) {
+                case 'matchdiff':
+                    types = {
+                        'DIGITMATCH': 1,
+                        'DIGITDIFF': 1
+                    };
+                    break;
+                case 'evenodd':
+                    types = {
+                        'DIGITEVEN': 1,
+                        'DIGITODD': 1
+                    };
+                    break;
+                case 'overunder':
+                    types = {
+                        'DIGITOVER': 1,
+                        'DIGITUNDER': 1
+                    };
+            }
+        }
+        for (var typeOfContract in types) {
+            if (types.hasOwnProperty(typeOfContract)) {
+                BinarySocket.send(Price.proposal(typeOfContract));
+            }
+        }
+    }
+
     return {
         proposal: createProposal,
         display: display,
@@ -288,7 +340,9 @@ var Price = (function() {
         incrFormId: function() {
             form_id++;
         },
-        clearFormId: clearFormId
+        clearFormId: clearFormId,
+        processForgetProposals: processForgetProposals,
+        processPriceRequest   : processPriceRequest,
     };
 
 })();
